move utilities.py into unit folder
