+++ testing/unit/test_sdf.py
@@ -2,7 +2,7 @@ import nanome
 import os
 
 from nanome.api import structure as struct
-from testing.utilities import assert_equal, assert_not_equal, TestOptions
+from testing.unit.utilities import assert_equal, assert_not_equal, TestOptions
 
 from nanome.util import Logs
 import unittest
