+++ testing/unit/test_ui.py
@@ -3,7 +3,7 @@ import nanome
 import nanome.api.ui as UI
 # from nanome.serialization.commands import ReceiveMenu, UpdateMenu
 from nanome._internal._ui._serialization import _LayoutNodeSerializer, _UIBaseSerializer
-from testing.utilities import (
+from testing.unit.utilities import (
     alter_object, assert_equal, rand_string, create_test, test_serializer, TestOptions)
 import unittest
